registration changes

+++ Multiserver/src/activitystreamer/util/ControlUtil.java
@@ -325,7 +325,7 @@ public class ControlUtil {
 			while (stringIterator.hasNext()){
 				String object = stringIterator.next();
 				if(null != object){
-					if(controlInstance.getLoad() > (((Integer)serverList.get(object).get("load")) + 2)) {
+					if(controlInstance.getLoad() > ((Long)serverList.get(object).get("load")).intValue() + 2) {
 						resultOutput.put("command", "REDIRECT");
 						resultOutput.put("hostname",(String) serverList.get(object).get("hostname"));
 						resultOutput.put("port",(String) serverList.get(object).get("port"));
